# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Stock Analysis Platform API
  version: 1.0.0
  description: RESTful API for stock analysis and portfolio management
  contact:
    name: API Support
    email: support@example.com
servers:
  -
    url: "http://localhost:3000"
    description: Development server
  -
    url: "https://api.example.com"
    description: Production server
paths:
...
  /api/settings:
    get:
      summary: Get current user's settings
      description: "Implements Requirements 7.2, 6.5: Restore and apply user's personalized settings (实现需求: 7.2, 6.5)"
      tags:
        - User Settings
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
    put:
      summary: Update current user's settings (partial update supported)
      description: "Implements Requirements 7.3, 6.5: Save settings changes to cloud in real-time (实现需求: 7.3, 6.5)"
      tags:
        - User Settings
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      id: 1
                      updatedAt: "2024-01-01T00:00:00Z"
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
  /api/reorder:
    put:
      summary: Reorder stocks in watchlist
      description: Implements Requirement 1.6
      tags:
        - Watchlist
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      id: 1
                      updatedAt: "2024-01-01T00:00:00Z"
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
  /api/:symbol/notes:
    patch:
      summary: Update notes for a watchlist item
      description: null
      tags:
        - Watchlist
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      id: 1
                      updatedAt: "2024-01-01T00:00:00Z"
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: symbol
          in: path
          description: symbol parameter
          required: true
          schema:
            type: string
  /api/:symbol/check:
    get:
      summary: Check if a stock is in watchlist
      description: null
      tags:
        - Watchlist
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: symbol
          in: path
          description: symbol parameter
          required: true
          schema:
            type: string
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...